refactoring continued
