+++ cpp/FMU-proxy/src/fmi/FmuHelper.cpp
@@ -28,7 +28,9 @@
 #include <cfloat>
 #include "FmiDefinitions.hpp"
 
-namespace fmuproxy::fmi {
+using namespace fmuproxy::fmi;
+
+namespace {
 
     IntegerAttribute toIntegerAttribute(fmi2_import_variable_t *variable, fmi2_xml_integer_variable_t *integer_variable) {
         IntegerAttribute attribute;
