+++ cpp/FMU-proxy/src/fmi/FmuInstance.hpp
@@ -28,63 +28,9 @@
 #include <iostream>
 #include <fmilib.h>
 #include "FmiSimulation.hpp"
+#include "FmuVariableAccess.hpp"
 
 namespace fmuproxy::fmi {
-
-    class VariableReader {
-
-    private:
-        fmi2_import_t *instance;
-        const fmi2_value_reference_t vr;
-
-    public:
-        VariableReader(fmi2_import_t *instance, fmi2_value_reference_t vr): instance(instance), vr(vr){}
-
-        fmi2_status_t readInteger(fmi2_integer_t &value) {
-            return fmi2_import_get_integer(instance, &vr, 1, &value);
-        }
-
-        fmi2_status_t readReal(fmi2_real_t &value) {
-            return fmi2_import_get_real(instance, &vr, 1, &value);
-        }
-
-        fmi2_status_t readString(fmi2_string_t &value) {
-            return fmi2_import_get_string(instance, &vr, 1, &value);
-        }
-
-        fmi2_status_t readBoolean(fmi2_boolean_t &value) {
-            return fmi2_import_get_boolean(instance, &vr, 1, &value);
-        }
-
-    };
-
-    class VariableWriter {
-
-    private:
-        fmi2_import_t *instance;
-        const fmi2_value_reference_t vr;
-
-    public:
-        VariableWriter(fmi2_import_t *instance, fmi2_value_reference_t vr): instance(instance), vr(vr){}
-
-        fmi2_status_t writeInteger(fmi2_integer_t value) {
-            return fmi2_import_set_integer(instance, &vr, 1, &value);
-        }
-
-        fmi2_status_t writeReal(fmi2_real_t value) {
-            return fmi2_import_set_real(instance, &vr, 1, &value);
-        }
-
-        fmi2_status_t writeString(fmi2_string_t value) {
-            return fmi2_import_set_string(instance, &vr, 1, &value);
-        }
-
-        fmi2_status_t writeBoolean(fmi2_boolean_t value) {
-            return fmi2_import_set_boolean(instance, &vr, 1, &value);
-        }
-
-    };
-
 
     class FmuInstance: FmiSimulation {
 
@@ -95,39 +41,40 @@ namespace fmuproxy::fmi {
         double current_time = 0.0;
         bool terminated = false;
 
-
     public:
         FmuInstance(fmi2_import_t* instance);
 
-        double getCurrentTime();
-
         bool isTerminated();
 
-        fmi2_value_reference_t get_value_reference(std::string name);
+        double getCurrentTime() override;
 
-        void init(double start, double stop);
+        void init(double start, double stop) override;
 
-        fmi2_status_t step(double step_size);
+        fmi2_status_t step(double step_size) override;
 
-        fmi2_status_t reset();
+        fmi2_status_t reset() override;
 
-        fmi2_status_t terminate();
+        fmi2_status_t terminate() override;
 
-        fmi2_status_t getInteger(fmi2_value_reference_t vr, fmi2_integer_t  &ref);
+        fmi2_status_t readInteger(fmi2_value_reference_t vr, fmi2_integer_t &ref) override;
+        fmi2_status_t readInteger(std::string name, fmi2_integer_t &ref);
 
-        fmi2_status_t getInteger(std::string name, fmi2_integer_t  &ref);
+        fmi2_status_t writeInteger(fmi2_value_reference_t vr, fmi2_integer_t value) override;
 
-        fmi2_status_t getReal(fmi2_value_reference_t vr, fmi2_real_t &ref);
+        fmi2_status_t readReal(fmi2_value_reference_t vr, fmi2_real_t &ref) override;
+        fmi2_status_t readReal(std::string name, fmi2_real_t &ref);
 
-        fmi2_status_t getReal(std::string name, fmi2_real_t &ref);
+        fmi2_status_t writeReal(fmi2_value_reference_t vr, fmi2_real_t value) override;
 
-        fmi2_status_t getString(fmi2_value_reference_t vr, fmi2_string_t &ref);
+        fmi2_status_t readString(fmi2_value_reference_t vr, fmi2_string_t &ref) override;
+        fmi2_status_t readString(std::string name, fmi2_string_t &ref);
 
-        fmi2_status_t getString(std::string name, fmi2_string_t &ref);
+        fmi2_status_t writeString(fmi2_value_reference_t vr, fmi2_string_t value) override;
 
-        fmi2_status_t getBoolean(fmi2_value_reference_t vr, fmi2_boolean_t &ref);
+        fmi2_status_t readBoolean(fmi2_value_reference_t vr, fmi2_boolean_t &ref) override;
+        fmi2_status_t readBoolean(std::string name, fmi2_boolean_t &ref);
 
-        fmi2_status_t getBoolean(std::string name, fmi2_boolean_t &ref);
+        fmi2_status_t writeBoolean(fmi2_value_reference_t vr, fmi2_boolean_t value) override;
 
         VariableReader getReader(fmi2_value_reference_t vr);
 
@@ -136,6 +83,8 @@ namespace fmuproxy::fmi {
         VariableWriter getWriter(fmi2_value_reference_t vr);
 
         VariableWriter getWriter(std::string name);
+
+        fmi2_value_reference_t get_value_reference(std::string name);
 
         ~FmuInstance();
