+++ cpp/FMU-proxy/src/test/TestUtil.hpp
@@ -7,14 +7,18 @@
 
 #include <iostream>
 
-std::string getOs() {
-#ifdef _WIN32
-    return "win32";
-#ielif _WIN64
-    return "win64";
-#elif __linux__
-    return "linux64";
-#endif
+namespace {
+
+    std::string getOs() {
+    #ifdef _WIN32
+            return "win32";
+    #ielif _WIN64
+        return "win64";
+    #elif __linux__
+            return "linux64";
+    #endif
+    }
+
 }
 
 #endif //FMU_PROXY_TESTUTIL_HPP
